Douge/guidance/edit (#26)
* Update our-journey-of-transforming-to-a-devops-culture.md

* Broad edit fixes and notes

Overall, it's an interesting topic that's too top-heavy with self-description, and a bit stingy with examples and clear recipes. I'd really tighten up the first sections, since they're a bit of a chore to read when you're wanting to get to the answer to the question: How, exactly, did you make it work? More examples and specific (short) use cases of what precisely you did in a certain situation would go a long way in payoff for the lengthy read, in my opinion. I'd refactor this and cut down on the scene-setting exposition at the top, and really dive into your examples, reciples, and techniques in the bottom 2/3rds. And short anecdotes! Everyone loves short anecdotes. :-)